files dir is optional

+++ env.sh
@@ -2,6 +2,8 @@
 # source env.sh scripts/dockerfile-1.txt files/dockerfile
 # source demorunner.sh scripts/dockerfile-1.txt
 
+# brew install coreutils (for greadlink)
+
 demo_script=""
 demo_files=""
 
@@ -10,22 +12,23 @@ if [ ! -f "${1}" ]; then
   kill -INT $$
 else
   demo_script="${1}"
+  demo_script_absolute_path=$(greadlink -f "${demo_script}")
+  demo_script_handle=$(echo $(basename "${demo_script}") | cut -d. -f1)
 fi
 
-if [ ! -d "${2}" ]; then
-  echo "Directory does not exist: [${2}]"
-  kill -INT $$
+if [ $# -gt 1 ]; then
+  if [ ! -d "${2}" ]; then
+    echo "Directory does not exist: [${2}]"
+    kill -INT $$
+  else
+    demo_files="${2}"
+    demo_files_absolute_path=$(greadlink -f "${demo_files}")
+  fi
 else
-  demo_files="${2}"
+  demo_files_absolute_path=""
 fi
 
 ##### DEMO ENV VARS
-
-# brew install coreutils (for greadlink)
-demo_script_absolute_path=$(greadlink -f "${demo_script}")
-demo_files_absolute_path=$(greadlink -f "${demo_files}")
-demo_script_handle=$(echo $(basename "${demo_script}") | cut -d. -f1)
-
 export DEMO_HOME=`pwd`
 export DEMO_SCRIPT="${demo_script_absolute_path}"
 export DEMO_FILES="${demo_files_absolute_path}"
@@ -100,7 +103,7 @@ fi
 # Stop running containers & prune images, containers, volumes, and networks (stopped,unused, and dangling)
 alias dclean="docker ps -a -q | xargs -n1 docker stop; docker system prune -af"
 # Remove all containers; prune dangling images; prune images, containers, volumes, and networks with specified label
-alias dclean2="docker ps -a -q | xargs -n1 docker rm -f; docker image prune -f; docker system prune -af --filter label=org.opencontainers.image.authors=me@example.com"
+alias dclean2="docker ps -a -q | xargs -n1 docker rm -f; docker image prune -f; docker system prune -af --filter label=com.example"
 
 # Rename Terminal tabs
 tabname() { printf '\e]1;%s\a' $1; }
